Rename app/ to renderer/
