+++ src/main/listenForKeyboardShortcut.ts
@@ -1,5 +1,5 @@
 import { BrowserWindow, ipcMain } from "electron"
-import { closeTheWindow } from "../app/ipcChannelNames"
+import { closeTheWindow } from "../renderer/ipcChannelNames"
 
 export const listenForKeyboardShortcutToCloseTheWindow: (window: BrowserWindow) => void = (
   window: BrowserWindow,
