+++ src/renderer/components/editor.tsx
