+++ src/renderer/components/styledEditorComponents.tsx
