+++ src/renderer/contentSecurityPolicy.ts
