+++ src/renderer/createTranscriptEditor.ts
