+++ src/renderer/file-drop-zone/basic-dropzone.tsx
